update component MiniNavigation.tsx
1. create hook useSelectClass.ts
