+++ Components/MiniNavigation/MiniNavigation.tsx
@@ -3,6 +3,7 @@ import classes from './MiniNavigation.module.css'
 import Link from 'next/link'
 import {RootState} from "../../redux/redusers/indexReduser";
 import {useSelector} from "react-redux";
+import {useSelectClass} from "./useSelectClass";
 
 type Props = {}
 const MiniNavigation = (props: Props) => {
@@ -14,27 +15,8 @@ const MiniNavigation = (props: Props) => {
     }
 
     const {dirName} = useSelector(selector)
-    const [classLink, setClassLink] = useState([classes.activeLink, '', '', '', ''])
+    const {classLink, classSelector} = useSelectClass(dirName, 'activeLink')
 
-    const classSelector = () => {
-        switch (dirName) {
-            case null :
-                setClassLink([classes.activeLink, '', '', '', ''])
-                break
-            case 'story' :
-                setClassLink(['', classes.activeLink, '', '', ''])
-                break
-            case 'society' :
-                setClassLink(['', '', classes.activeLink, '', ''])
-                break
-            case 'jurisprudence' :
-                setClassLink(['', '', '', classes.activeLink, ''])
-                break
-            case 'economy' :
-                setClassLink(['', '', '', '', classes.activeLink])
-                break
-        }
-    }
 
     useEffect(()=> {
         classSelector()
@@ -47,31 +29,31 @@ const MiniNavigation = (props: Props) => {
                 <nav>
                     <ul>
                         <Link href={'/'}>
-                            <a className={classLink[0]}>
+                            <a className={classes[classLink[0]]}>
                                 <li>Главная</li>
                                 <span/>
                             </a>
                         </Link>
                         <Link href={'/post/story?id=0'}>
-                            <a className={classLink[1]}>
+                            <a className={classes[classLink[1]]}>
                                 <li>История</li>
                                 <span/>
                             </a>
                         </Link>
                         <Link href={'/post/society?id=0'}>
-                            <a className={classLink[2]}>
+                            <a className={classes[classLink[2]]}>
                                 <li>Общество</li>
                                 <span/>
                             </a>
                         </Link>
                         <Link href={'/post/jurisprudence?id=0'}>
-                            <a className={classLink[3]}>
+                            <a className={classes[classLink[3]]}>
                                 <li>Право</li>
                                 <span/>
                             </a>
                         </Link>
                         <Link href={'/post/economy?id=0'}>
-                            <a className={classLink[4]}>
+                            <a className={classes[classLink[4]]}>
                                 <li>Экономика</li>
                                 <span/>
                             </a>
